Separate out legacy code.
