+++ src/tracer/BUILD
@@ -20,26 +20,6 @@ lightstep_cc_library(
 )
 
 lightstep_cc_library(
-    name = "propagation_lib",
-    private_hdrs = [
-        "propagation.h",
-    ],
-    srcs = [
-        "propagation.cpp",
-    ],
-    deps = [
-        "//src/common:utility_lib",
-        "//src/common:in_memory_stream_lib",
-        "//3rd_party/base64:base64_lib",
-        "//lightstep-tracer-common:lightstep_carrier_proto_cc",
-    ],
-    external_deps = [
-        "@com_google_protobuf//:protobuf",
-        "@io_opentracing_cpp//:opentracing",
-    ],
-)
-
-lightstep_cc_library(
     name = "binary_carrier_lib",
     srcs = [
         "binary_carrier.cpp",
@@ -70,20 +50,7 @@ lightstep_cc_library(
         "lightstep_span_context.cpp",
     ],
     deps = [
-        ":propagation_lib",
-    ],
-)
-
-lightstep_cc_library(
-    name = "lightstep_immutable_span_context_lib",
-    private_hdrs = [
-        "lightstep_immutable_span_context.h",
-    ],
-    srcs = [
-        "lightstep_immutable_span_context.cpp",
-    ],
-    deps = [
-        ":lightstep_span_context_interface",
+        "//src/tracer/legacy:propagation_lib",
     ],
 )
 
@@ -109,23 +76,6 @@ lightstep_cc_library(
 )
 
 lightstep_cc_library(
-    name = "lightstep_span_lib",
-    private_hdrs = [
-        "lightstep_span.h",
-    ],
-    srcs = [
-        "lightstep_span.cpp",
-    ],
-    deps = [
-        "//src/common:logger_lib",
-        "//src/common:utility_lib",
-        "//src/common:random_lib",
-        "//src/recorder:recorder_interface",
-        ":lightstep_span_context_interface",
-    ],
-)
-
-lightstep_cc_library(
     name = "tracer_impl_lib",
     private_hdrs = [
         "tracer_impl.h",
@@ -136,29 +86,12 @@ lightstep_cc_library(
     deps = [
         "//src/common:logger_lib",
         "//src/recorder:recorder_interface",
-        ":propagation_lib",
-        ":lightstep_span_lib",
-        ":lightstep_immutable_span_context_lib",
+        "//src/tracer/legacy:lightstep_span_lib",
+        "//src/tracer/legacy:propagation_lib",
+        "//src/tracer/legacy:lightstep_immutable_span_context_lib",
     ],
 )
 
-
-lightstep_cc_library(
-    name = "lightstep_tracer_impl_lib",
-    private_hdrs = [
-        "lightstep_tracer_impl.h",
-    ],
-    srcs = [
-        "lightstep_tracer_impl.cpp",
-    ],
-    deps = [
-        "//src/common:logger_lib",
-        "//src/recorder:recorder_interface",
-        ":propagation_lib",
-        ":lightstep_span_lib",
-        ":lightstep_immutable_span_context_lib",
-    ],
-)
 
 lightstep_cc_library(
     name = "lightstep_tracer_factory_lib",
@@ -187,9 +120,9 @@ lightstep_cc_library(
         "//src/recorder:grpc_transporter_interface",
         "//src/recorder:manual_recorder_lib",
         "//src/recorder:stream_recorder_interface",
+        "//src/tracer/legacy:lightstep_tracer_impl_lib",
         "//lightstep-tracer-common:collector_proto_cc",
         "//:config_lib",
-        ":lightstep_tracer_impl_lib",
     ],
     external_deps = [
         "@io_opentracing_cpp//:opentracing",
