+++ src/tracer/legacy/lightstep_tracer_impl.cpp
@@ -1,6 +1,6 @@
-#include "tracer/lightstep_tracer_impl.h"
-#include "tracer/lightstep_immutable_span_context.h"
-#include "tracer/lightstep_span.h"
+#include "tracer/legacy/lightstep_tracer_impl.h"
+#include "tracer/legacy/lightstep_immutable_span_context.h"
+#include "tracer/legacy/lightstep_span.h"
 
 namespace lightstep {
 const auto DefaultFlushTimeout = std::chrono::seconds{10};
