+++ src/tracer/legacy/lightstep_tracer_impl.h
@@ -4,7 +4,7 @@
 
 #include "common/logger.h"
 #include "recorder/recorder.h"
-#include "tracer/propagation.h"
+#include "tracer/legacy/propagation.h"
 
 namespace lightstep {
 class LightStepTracerImpl final
