+++ src/tracer/tracer_impl.cpp
@@ -1,6 +1,6 @@
 #include "tracer/tracer_impl.h"
 
-#include "tracer/lightstep_immutable_span_context.h"
+#include "tracer/legacy/lightstep_immutable_span_context.h"
 
 namespace lightstep {
 const auto DefaultFlushTimeout = std::chrono::seconds{10};
