+++ test/recorder/auto_recorder_test.cpp
@@ -7,7 +7,7 @@
 #include "test/recorder/in_memory_sync_transporter.h"
 #include "test/testing_condition_variable_wrapper.h"
 #include "test/utility.h"
-#include "tracer/lightstep_tracer_impl.h"
+#include "tracer/legacy/lightstep_tracer_impl.h"
 
 using namespace lightstep;
 using namespace opentracing;
