+++ test/recorder/stream_recorder/BUILD
@@ -13,7 +13,7 @@ lightstep_catch_test(
         "stream_recorder_test.cpp",
     ],
     deps = [
-        "//src/tracer:lightstep_tracer_impl_lib",
+        "//src/tracer/legacy:lightstep_tracer_impl_lib",
         "//src/recorder/stream_recorder:stream_recorder_lib",
         "//src/network/ares_dns_resolver:ares_dns_resolver_lib",
         "//test/mock_satellite:mock_satellite_lib",
